# apps/auth-server/api-docs.json
{"openapi":"3.0.0","paths":{"/callback":{"get":{"operationId":"AppController_callback","parameters":[],"responses":{"200":{"description":""}}}},"/":{"get":{"operationId":"AppController_index","parameters":[],"responses":{"200":{"description":""}}}},"/oidc/*":{"all":{"operationId":"OidcController_mountedOidc","parameters":[],"responses":{"200":{"description":""}}}},"/interaction/{uid}":{"get":{"operationId":"InteractionController_login","parameters":[{"name":"uid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"post":{"operationId":"InteractionController_loginCheck","parameters":[{"name":"uid","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}},"/interaction/{uid}/confirm":{"post":{"operationId":"InteractionController_confirmLogin","parameters":[{"name":"uid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/interaction/{uid}/abort":{"get":{"operationId":"InteractionController_abortLogin","parameters":[{"name":"uid","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/ldap":{"post":{"operationId":"LdapController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"201":{"description":"Register user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResultDto"}}}}}},"patch":{"operationId":"LdapController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":"Update or change user password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResultDto"}}}}}},"delete":{"operationId":"LdapController_deletion","parameters":[],"responses":{"200":{"description":"Delete user","content":{"application/json":{"schema":{"type":"number"}}}}}},"get":{"operationId":"LdapController_findByEmail","parameters":[{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Find user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"info":{"title":"Authentication server","description":"Authentication with oidc provider documentation","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3333"}],"components":{"securitySchemes":{"oidc":{"type":"openIdConnect","description":"Open id server","name":"oidc","openIdConnectUrl":"http://localhost:3333/undefined/.well-known/openid-configuration"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RegisterUserDto":{"type":"object","properties":{"email":{"type":"string"},"phoneNumber":{"type":"string"},"name":{"type":"string"},"familyName":{"type":"string"},"birthdate":{"format":"date-time","type":"string"},"gender":{"type":"string","enum":["man","woman","other"]},"password":{"type":"string"},"confirmationPassword":{"type":"string"}},"required":["email","phoneNumber","name","familyName","birthdate","gender","confirmationPassword"]},"UserDto":{"type":"object","properties":{"email":{"type":"string"},"phoneNumber":{"type":"string"},"name":{"type":"string"},"familyName":{"type":"string"},"birthdate":{"format":"date-time","type":"string"},"gender":{"type":"string","enum":["man","woman","other"]},"password":{"type":"string"}},"required":["email","phoneNumber","name","familyName","birthdate","gender"]},"RegisterResultDto":{"type":"object","properties":{"message":{"type":"string"},"user":{"description":"User included in result","allOf":[{"$ref":"#/components/schemas/UserDto"}]}},"required":["message","user"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"phoneNumber":{"type":"string"},"name":{"type":"string"},"familyName":{"type":"string"},"birthdate":{"format":"date-time","type":"string"},"gender":{"type":"string","enum":["man","woman","other"]},"password":{"type":"string"},"confirmationPassword":{"type":"string"},"oldPassword":{"type":"string"}},"required":["email","phoneNumber","name","familyName","birthdate","gender"]},"UpdateResultDto":{"type":"object","properties":{"message":{"type":"string"},"result":{"type":"array","items":{"type":"string"}}},"required":["message","result"]}}}}